handle Network.responseReceived w/o requestHeaders
Newer chromium is not including requestHeaders in
Network.responseReceived messages.
https://stackoverflow.com/questions/55842059/how-to-get-the-request-headers-using-the-chrome-devtool-protocol
Umbra was getting headers from Network.responseReceived instead of
Network.requestWillBeSent "because it has request info that we
need, like cookies" according to the commit message, see
https://github.com/internetarchive/umbra/commit/063a58d8539

We could switch back to Network.requestWillBeSent to be able to use a
subset of the headers. As of now I don't see a good way to get the
cookies though. There must be a way because you can still see cookies in
the chrome dev tools UI. But how much dev work does this warrant, when
brozzler is an option :)